Added enum for toggle between how we offload cache after the forward pass
By default we count the layerwise executions in the forward pass
but in the case of using cuda graphs with trtllm, forward passes
two and greater don`t trigger our kvbm counts. Thus, we hook into
wait_for_save() which is called at the end of every forward pass,
bypassing the need to count the layers from the stream events.

wait_for_save() might be slower since when getting the stream of events
we only have to sync on the cuda events to know it is safe to
transfer. This can be optimized later.